records: Move TestPID_UnmarshalJSON to state package
